Extract detection row components for reuse across tabs
AdReviewTab was 527 lines and its row, card, badge, and action internals carry
non-obvious responsive layout constraints in comments. The Detected Ads tab
needs exactly those, so they move to DetectionRows rather than being copied.

Actions are now supplied per view: each renders only when its handler is passed,
and Confirm and Not an ad additionally require an unresolved row. Category and
action chips are opt-in, since they are noise on the review queue and the point
of the Detected Ads tab.

Pure refactor. The 19 existing AdReviewTab tests pass unmodified.
